+++ test/Taxonomy/Services/ServiceTest.php
@@ -1,0 +1,68 @@
+<?php
+namespace DTS\eBaySDK\Test\Taxonomy\Services;
+
+use DTS\eBaySDK\Taxonomy\Services\TaxonomyBaseService;
+use DTS\eBaySDK\Taxonomy\Services\TaxonomyService;
+use DTS\eBaySDK\Test\Taxonomy\Mocks\Service;
+use DTS\eBaySDK\Test\Mocks\HttpRestHandler;
+
+class ServiceTest extends \PHPUnit_Framework_TestCase
+{
+    public function testConfigDefinitions()
+    {
+        $d = TaxonomyBaseService::getConfigDefinitions();
+
+        $this->assertArrayHasKey('apiVersion', $d);
+        $this->assertEquals([
+            'valid' => ['string'],
+            'default' => TaxonomyService::API_VERSION,
+            'required' => true
+        ], $d['apiVersion']);
+
+        $this->assertArrayHasKey('authorization', $d);
+        $this->assertEquals([
+            'valid'   => ['string'],
+            'required' => true
+        ], $d['authorization']);
+
+        $this->assertArrayHasKey('marketplaceId', $d);
+        $this->assertEquals([
+            'valid' => ['string']
+        ], $d['marketplaceId']);
+    }
+
+    public function testRequiredEbayHeaders()
+    {
+        $h = new HttpRestHandler();
+
+        $s = new Service([
+            'authorization' => '321',
+            'httpHandler' => $h
+        ]);
+
+        $s->testOperation();
+
+        // Test required headers first.
+        $this->assertArrayHasKey(TaxonomyBaseService::HDR_AUTHORIZATION, $h->headers);
+        $this->assertEquals('Bearer 321', $h->headers[TaxonomyBaseService::HDR_AUTHORIZATION]);
+
+        // Test that optional headers have not been set until they have been configured.
+        $this->assertArrayNotHasKey(TaxonomyBaseService::HDR_MARKETPLACE_ID, $h->headers);
+    }
+
+    public function testOptionalEbayHeaders()
+    {
+        $h = new HttpRestHandler();
+
+        $s = new Service([
+            'authorization' => '321',
+            'marketplaceId' => '123',
+            'httpHandler' => $h
+        ]);
+
+        $s->testOperation();
+
+        $this->assertArrayHasKey(TaxonomyBaseService::HDR_MARKETPLACE_ID, $h->headers);
+        $this->assertEquals('123', $h->headers[TaxonomyBaseService::HDR_MARKETPLACE_ID]);
+    }
+}